# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nepal Summer League Finland 2026 Tournament API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AuthResponse {
  isAdmin: boolean;
  username: string;
}

export interface Team {
  id: number;
  name: string;
  shortName: string;
  primaryColor: string;
  /** @nullable */
  logoUrl?: string | null;
  createdAt?: string;
}

export interface TeamInput {
  name: string;
  shortName: string;
  primaryColor: string;
  /** @nullable */
  logoUrl?: string | null;
}

export interface TeamUpdate {
  name?: string;
  shortName?: string;
  primaryColor?: string;
  /** @nullable */
  logoUrl?: string | null;
}

export type MatchStatus = typeof MatchStatus[keyof typeof MatchStatus];


export const MatchStatus = {
  upcoming: 'upcoming',
  live: 'live',
  finished: 'finished',
} as const;

export interface Match {
  id: number;
  matchNumber: number;
  homeTeamId: number;
  awayTeamId: number;
  /** @nullable */
  homeTeamName?: string | null;
  /** @nullable */
  awayTeamName?: string | null;
  /** @nullable */
  homeTeamShortName?: string | null;
  /** @nullable */
  awayTeamShortName?: string | null;
  homeScore?: number;
  awayScore?: number;
  scheduledTime: string;
  pitch: number;
  status: MatchStatus;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  finishedAt?: string | null;
  createdAt?: string;
}

export type MatchDetailStatus = typeof MatchDetailStatus[keyof typeof MatchDetailStatus];


export const MatchDetailStatus = {
  upcoming: 'upcoming',
  live: 'live',
  finished: 'finished',
} as const;

export interface Goal {
  id: number;
  matchId: number;
  teamId: number;
  /** @nullable */
  teamName?: string | null;
  /** @nullable */
  scorerName?: string | null;
  /** @nullable */
  assistName?: string | null;
  minute: number;
  isOwnGoal?: boolean;
  createdAt?: string;
}

export type CardCardType = typeof CardCardType[keyof typeof CardCardType];


export const CardCardType = {
  yellow: 'yellow',
  red: 'red',
} as const;

export interface Card {
  id: number;
  matchId: number;
  teamId: number;
  /** @nullable */
  teamName?: string | null;
  /** @nullable */
  playerName?: string | null;
  cardType: CardCardType;
  minute: number;
  createdAt?: string;
}

export type MatchEventEventType = typeof MatchEventEventType[keyof typeof MatchEventEventType];


export const MatchEventEventType = {
  goal: 'goal',
  own_goal: 'own_goal',
  yellow_card: 'yellow_card',
  red_card: 'red_card',
  substitution: 'substitution',
  kickoff: 'kickoff',
  halftime: 'halftime',
  fulltime: 'fulltime',
  var_review: 'var_review',
  injury: 'injury',
} as const;

export interface MatchEvent {
  id: number;
  matchId: number;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  teamName?: string | null;
  eventType: MatchEventEventType;
  minute: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  playerName?: string | null;
  createdAt?: string;
}

export interface MatchDetail {
  id: number;
  matchNumber: number;
  homeTeamId: number;
  awayTeamId: number;
  /** @nullable */
  homeTeamName?: string | null;
  /** @nullable */
  awayTeamName?: string | null;
  /** @nullable */
  homeTeamShortName?: string | null;
  /** @nullable */
  awayTeamShortName?: string | null;
  homeScore?: number;
  awayScore?: number;
  scheduledTime: string;
  pitch: number;
  status: MatchDetailStatus;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  finishedAt?: string | null;
  createdAt?: string;
  goals: Goal[];
  cards: Card[];
  events: MatchEvent[];
}

export interface MatchInput {
  homeTeamId: number;
  awayTeamId: number;
  scheduledTime: string;
  pitch: number;
  matchNumber: number;
}

export type MatchUpdateStatus = typeof MatchUpdateStatus[keyof typeof MatchUpdateStatus];


export const MatchUpdateStatus = {
  upcoming: 'upcoming',
  live: 'live',
  finished: 'finished',
} as const;

export interface MatchUpdate {
  homeScore?: number;
  awayScore?: number;
  scheduledTime?: string;
  pitch?: number;
  status?: MatchUpdateStatus;
}

export interface GoalInput {
  teamId: number;
  /** @nullable */
  scorerName?: string | null;
  /** @nullable */
  assistName?: string | null;
  minute: number;
  isOwnGoal?: boolean;
}

export type CardInputCardType = typeof CardInputCardType[keyof typeof CardInputCardType];


export const CardInputCardType = {
  yellow: 'yellow',
  red: 'red',
} as const;

export interface CardInput {
  teamId: number;
  /** @nullable */
  playerName?: string | null;
  cardType: CardInputCardType;
  minute: number;
}

export type MatchEventInputEventType = typeof MatchEventInputEventType[keyof typeof MatchEventInputEventType];


export const MatchEventInputEventType = {
  goal: 'goal',
  own_goal: 'own_goal',
  yellow_card: 'yellow_card',
  red_card: 'red_card',
  substitution: 'substitution',
  kickoff: 'kickoff',
  halftime: 'halftime',
  fulltime: 'fulltime',
  var_review: 'var_review',
  injury: 'injury',
} as const;

export interface MatchEventInput {
  /** @nullable */
  teamId?: number | null;
  eventType: MatchEventInputEventType;
  minute: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  playerName?: string | null;
}

export interface StandingRow {
  position: number;
  teamId: number;
  teamName: string;
  teamShortName?: string;
  primaryColor?: string;
  played: number;
  won: number;
  drawn: number;
  lost: number;
  goalsFor: number;
  goalsAgainst: number;
  goalDifference: number;
  points: number;
}

export interface TopScorer {
  scorerName: string;
  teamId: number;
  teamName: string;
  goals: number;
}

/**
 * @nullable
 */
export type TournamentStatsHighestScoringMatch = {
  matchId?: number;
  homeTeam?: string;
  awayTeam?: string;
  homeScore?: number;
  awayScore?: number;
  totalGoals?: number;
} | null;

/**
 * @nullable
 */
export type TournamentStatsMostGoalsTeam = {
  teamId?: number;
  teamName?: string;
  goals?: number;
} | null;

/**
 * @nullable
 */
export type TournamentStatsBestDefenseTeam = {
  teamId?: number;
  teamName?: string;
  goalsAgainst?: number;
} | null;

/**
 * @nullable
 */
export type TournamentStatsMostWinsTeam = {
  teamId?: number;
  teamName?: string;
  wins?: number;
} | null;

export interface TournamentStats {
  totalMatches: number;
  matchesPlayed: number;
  liveMatches?: number;
  totalGoals: number;
  averageGoalsPerMatch: number;
  /** @nullable */
  highestScoringMatch: TournamentStatsHighestScoringMatch;
  /** @nullable */
  mostGoalsTeam?: TournamentStatsMostGoalsTeam;
  /** @nullable */
  bestDefenseTeam?: TournamentStatsBestDefenseTeam;
  /** @nullable */
  mostWinsTeam?: TournamentStatsMostWinsTeam;
}

export type ListMatchesParams = {
status?: ListMatchesStatus;
};

export type ListMatchesStatus = typeof ListMatchesStatus[keyof typeof ListMatchesStatus];


export const ListMatchesStatus = {
  upcoming: 'upcoming',
  live: 'live',
  finished: 'finished',
} as const;