# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nepal Summer League Finland 2026 Tournament API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getAdminMe
      tags: [auth]
      summary: Get current admin session
      responses:
        "200":
          description: Current admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Not authenticated

  # ── Teams ─────────────────────────────────────────────────────────────
...
  /matches/{matchId}/goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List goals for a match
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Goals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: addGoal
      tags: [goals]
      summary: Add a goal to a match
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Goal added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /matches/{matchId}/goals/{goalId}:
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
        - name: goalId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Cards ─────────────────────────────────────────────────────────────
  /matches/{matchId}/cards:
    get:
      operationId: listCards
      tags: [cards]
      summary: List cards for a match
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cards list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
    post:
      operationId: addCard
      tags: [cards]
      summary: Add a card to a match
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardInput"
      responses:
        "201":
          description: Card added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /matches/{matchId}/cards/{cardId}:
    delete:
      operationId: deleteCard
      tags: [cards]
      summary: Delete a card
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Events ────────────────────────────────────────────────────────────
  /matches/{matchId}/events:
    get:
      operationId: listMatchEvents
      tags: [events]
      summary: List timeline events for a match
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Events list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchEvent"
    post:
      operationId: addMatchEvent
      tags: [events]
      summary: Add a timeline event to a match
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchEventInput"
      responses:
        "201":
          description: Event added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchEvent"

  # ── Standings ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [isAdmin, username]
      properties:
        isAdmin:
          type: boolean
        username:
          type: string

    # ── Team ─────────────────────────────────────────────────────────
...
    Goal:
      type: object
      required: [id, matchId, teamId, minute]
      properties:
        id:
          type: integer
        matchId:
          type: integer
        teamId:
          type: integer
        teamName:
          type: ["string", "null"]
        scorerName:
          type: ["string", "null"]
        assistName:
          type: ["string", "null"]
        minute:
          type: integer
        isOwnGoal:
          type: boolean
        createdAt:
          type: string
    GoalInput:
      type: object
      required: [teamId, minute]
      properties:
        teamId:
          type: integer
        scorerName:
          type: ["string", "null"]
        assistName:
          type: ["string", "null"]
        minute:
          type: integer
        isOwnGoal:
          type: boolean

    # ── Card ─────────────────────────────────────────────────────────
    Card:
      type: object
      required: [id, matchId, teamId, cardType, minute]
      properties:
        id:
          type: integer
        matchId:
          type: integer
        teamId:
          type: integer
        teamName:
          type: ["string", "null"]
        playerName:
          type: ["string", "null"]
        cardType:
          type: string
          enum: [yellow, red]
        minute:
          type: integer
        createdAt:
          type: string
    CardInput:
      type: object
      required: [teamId, cardType, minute]
      properties:
        teamId:
          type: integer
        playerName:
          type: ["string", "null"]
        cardType:
          type: string
          enum: [yellow, red]
        minute:
          type: integer

    # ── Match Event ───────────────────────────────────────────────────
    MatchEvent:
      type: object
      required: [id, matchId, eventType, minute]
      properties:
        id:
          type: integer
        matchId:
          type: integer
        teamId:
          type: ["integer", "null"]
        teamName:
          type: ["string", "null"]
        eventType:
          type: string
          enum: [goal, own_goal, yellow_card, red_card, substitution, kickoff, halftime, fulltime, var_review, injury]
        minute:
          type: integer
        description:
          type: ["string", "null"]
        playerName:
          type: ["string", "null"]
        createdAt:
          type: string
    MatchEventInput:
      type: object
      required: [eventType, minute]
      properties:
        teamId:
          type: ["integer", "null"]
        eventType:
          type: string
          enum: [goal, own_goal, yellow_card, red_card, substitution, kickoff, halftime, fulltime, var_review, injury]
        minute:
          type: integer
        description:
          type: ["string", "null"]
        playerName:
          type: ["string", "null"]

    # ── Standings ─────────────────────────────────────────────────────
...